+++ ui/pages/create-account/connect-hardware/account-list.test.tsx
@@ -5,7 +5,11 @@ import configureStore from '../../../store/store';
 import mockState from '../../../../test/data/mock-state.json';
 import { enLocale as messages } from '../../../../test/lib/i18n-helpers';
 import AccountList from './account-list';
-import { LATTICE_HD_PATHS, LEDGER_HD_PATHS, TREZOR_HD_PATHS } from '.';
+import {
+  LATTICE_HD_PATHS,
+  LEDGER_HD_PATHS,
+  TREZOR_HD_PATHS,
+} from './utils/hardware-hd-paths';
 
 const render = () => {
   const store = configureStore({
